+++ iOS/issue-tracker/issue-tracker/Protocol/Identyfying.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+protocol Identifying {
+    static var identifier: String { get }
+}
+
+extension Identifying {
+    static var identifier: String {
+        return String(describing: Self.self)
+    }
+}
