+++ iOS/issue-tracker/issue-tracker/Protocol/JWTManageable.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol JWTManageable {
+    func get() -> String?
+    func set(jwt: String) -> Bool
+}
