+++ iOS/issue-tracker/issue-tracker/Protocol/URLSessionProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol URLSessionProtocol {
+    func dataTaskPublisher(for request: URLRequest) -> URLSession.DataTaskPublisher
+}
+
+extension URLSession: URLSessionProtocol { }
