+++ iOS/issue-tracker/issue-tracker/View/IssueTableViewCell.swift
@@ -1,15 +1,5 @@
 import UIKit
 
-protocol Identifying {
-    static var identifier: String { get }
-}
-
-extension Identifying {
-    static var identifier: String {
-        return String(describing: Self.self)
-    }
-}
-
 class IssueTableViewCell: UITableViewCell, Identifying {
     
     static var nib: UINib {
