+++ iOS/issue-tracker/issue-tracker/ViewController/IssueAdditionViewController.swift
@@ -15,5 +15,3 @@ class IssueAdditionViewController: UIViewController {
     }
     
 }
-
-extension IssueAdditionViewController: Identifying { }
