+++ iOS/issue-tracker/issue-tracker/ViewController/IssueListViewController.swift
@@ -136,5 +136,3 @@ extension IssueListViewController: UITableViewDataSource, UITableViewDelegate {
     }
     
 }
-
-extension IssueListViewController: Identifying { }
