+++ iOS/issue-tracker/issue-trackerTests/JWTManagerTest.swift
@@ -32,11 +32,4 @@ class JWTManagerTest: XCTestCase {
         XCTAssertNotEqual(get, "234")
     }
     
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-
 }
