+++ iOS/issue-tracker/issue-trackerTests/Mock/JWTManagerStub.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+class JWTManagerStub: JWTManageable {
+    
+    var string: String?
+    
+    func get() -> String? {
+        return string
+    }
+    
+    func set(jwt: String) -> Bool {
+        return true
+    }
+    
+}
