+++ iOS/issue-tracker/issue-trackerTests/Mock/TestDTO.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct TestDTO: Decodable {
+    let test: String
+}
