+++ iOS/issue-tracker/issue-trackerTests/Mock/URLSessionStub.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+class URLSessionStub: URLSessionProtocol {
+    
+    var request: URLRequest?
+    
+    func dataTaskPublisher(for request: URLRequest) -> URLSession.DataTaskPublisher {
+        self.request = request
+        return URLSession.DataTaskPublisher(request: URLRequest(url: URL(string: "a")!), session: URLSession.init(configuration: .default))
+    }
+    
+}
