+++ iOS/issue-tracker/issue-trackerTests/NetworkManagerTest.swift
@@ -37,33 +37,5 @@ class NetworkManagerTest: XCTestCase {
         XCTAssertEqual(urlSessionStub.request?.url?.absoluteString, expectedAddress)
         XCTAssertEqual(urlSessionStub.request?.value(forHTTPHeaderField: "Authorization"), expectedHeaderValue)
     }
-}
-
-class URLSessionStub: URLSessionProtocol {
     
-    var request: URLRequest?
-    
-    func dataTaskPublisher(for request: URLRequest) -> URLSession.DataTaskPublisher {
-        self.request = request
-        return URLSession.DataTaskPublisher(request: URLRequest(url: URL(string: "a")!), session: URLSession.init(configuration: .default))
-    }
-    
-}
-
-class JWTManagerStub: JWTManageable {
-    
-    var string: String?
-    
-    func get() -> String? {
-        return string
-    }
-    
-    func set(jwt: String) -> Bool {
-        return true
-    }
-    
-}
-
-struct TestDTO: Decodable {
-    let test: String
 }
