+++ iOS/issue-tracker/issue-trackerTests/issue_trackerTests.swift
@@ -1,33 +1,6 @@
-//
-//  issue_trackerTests.swift
-//  issue-trackerTests
-//
-//  Created by sonjuhyeong on 2021/06/07.
-//
-
 import XCTest
 @testable import issue_tracker
 
 class IssueTrackerTests: XCTestCase {
-
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() throws {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
 
 }
